+++ Sources/Nimble/AsyncExpression.swift
@@ -120,4 +120,3 @@ public struct AsyncExpression<Value> {
         )
     }
 }
-
